PSG-2073 pull user types into own file in /types, export WebAuthnDevices, UserInfo, UpdateUserPayload and CreateUserPayload
