+++ src/classes/User.ts
@@ -2,71 +2,7 @@
 import { PassageConfig } from '../types/PassageConfig';
 import axios from '../utils/axios';
 import { PassageError } from './PassageError';
-
-interface UserEventInfo {
-  type: string;
-  timestamp: string;
-  id: string;
-  ip_addr: string;
-  user_agent: string;
-}
-
-interface WebAuthnDevices {
-  id: string;
-  cred_id: string;
-  friendly_name: string;
-  usage_count: number;
-  updated_at: string;
-  created_at: string;
-  last_login_at: string;
-}
-
-enum UserStatus {
-  ACTIVE = 'active',
-  INACTIVE = 'inactive',
-  PENDING = 'pending',
-}
-type UserStatusEnum = keyof typeof UserStatus;
-
-interface PossibleUserUpdateAttributes {
-  email: string;
-  phone: string;
-}
-
-type UserAttributes = Pick<PossibleUserUpdateAttributes, 'email'> | Pick<PossibleUserUpdateAttributes, 'phone'>;
-
-export interface Metadata {
-  [key: string]: boolean | string | number;
-}
-
-interface UserObject {
-  created_at: string;
-  updated_at: string;
-  status: UserStatusEnum;
-  email_verified: boolean;
-  phone_verified: boolean;
-  email: string;
-  phone: string;
-  id: string;
-  last_login_at: string;
-  login_count: number;
-  recent_events: Array<UserEventInfo>;
-  webauthn: boolean;
-  webauthn_devices: Array<WebAuthnDevices>;
-  user_metadata?: Metadata;
-}
-
-interface UpdateUserPayload {
-  email?: string;
-  phone?: string;
-  user_metadata?: Metadata;
-}
-
-interface CreateUserPayload {
-  email?: string;
-  phone?: string;
-  user_metadata?: Metadata;
-}
+import { WebAuthnDevices, UserObject, UpdateUserPayload, CreateUserPayload } from '../types/User';
 
 /***/
 export default class User {
